requirements.txt
python-dateutil==2.8.2
urllib3==1.26.12
slack_sdk==3.19.2

requests~=2.28.1
validators~=0.20.0